requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
streamlit==1.39.0
joblib==1.4.2
matplotlib==3.9.2
PyYAML==6.0.2
pandera==0.20.4
mlflow==2.17.0
shap==0.46.0
pytest==8.3.3
ruff==0.6.9
